+++ src/assets/content.css
@@ -25,11 +25,6 @@
   margin-bottom: 20px;
 }
 
-.menu-section {
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 20px;
-}
 
 .input-group {
   display: flex;
@@ -45,26 +40,6 @@ input[type="text"] {
   border: 1px solid #ccc;
   border-radius: 4px;
   margin-right: 10px;
-}
-
-input[type="number"] {
-  flex: 1;
-  padding: 10px;
-  font-size: 16px;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-  margin-right: 10px;
-  width: 75%;
-}
-
-input[type="number"]::-webkit-outer-spin-button,
-input[type="number"]::-webkit-inner-spin-button {
-  -webkit-appearance: none;
-  margin: 0;
-}
-
-input[type="number"] {
-  -moz-appearance: textfield;
 }
 
 .input-and-button {
@@ -115,35 +90,10 @@ button:hover {
   position: relative; /* close 버튼 위치 조정을 위해 추가 */
 }
 
-.menu-card {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  padding: 10px;
-  margin: 5px;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-  background-color: #f8f9fa;
-  width: 267px;
-}
-
 .attendee-name {
   font-size: 16px;
   text-align: center; /* 텍스트 가운데 정렬 */
   margin: 0;
-}
-
-.menu-list {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: flex-end;
-}
-
-.item-name,
-.item-price,
-.item-quantity {
-  font-size: 16px;
 }
 
 .close {
@@ -162,34 +112,8 @@ button:hover {
   right: 10px; /* 카드의 오른쪽으로부터 10px 떨어진 위치 */
 }
 
-.menu-close {
-  color: #aaa;
-  border: none;
-  border-radius: 50%;
-  width: 20px;
-  height: 20px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  cursor: pointer;
-  font-size: 25px;
-  top: 10px; /* 카드의 위쪽으로부터 10px 떨어진 위치 */
-  right: 10px; /* 카드의 오른쪽으로부터 10px 떨어진 위치 */
-}
-
-.close .menu-close:hover {
+.close {
   color: black;
-}
-
-.total-price {
-  display: flex;
-  justify-content: center;
-  margin-bottom: 1rem;
-  font-size: 25px;
-  font-family: "Gamja Flower", sans-serif;
-  font-weight: 400;
-  font-style: normal;
-  color: rgb(70, 70, 189);
 }
 
 .attendee-total-price {
